move validation tests into dedicated files
